fix : header problem
헤더파일 순환참조 문제 해결했습니다.
