+++ pintos/include/vm/vm.h
@@ -1,33 +1,15 @@
 #ifndef VM_VM_H
 #define VM_VM_H
-#include <stdbool.h>
+
 #include <hash.h>
+#include <stdbool.h>
 #include "threads/palloc.h"
+#include "vm/vm_type.h"
 
-enum vm_type {
-    /* page not initialized */
-    VM_UNINIT = 0,
-    /* page not related to the file, aka anonymous page */
-    VM_ANON = 1,
-    /* page that realated to the file */
-    VM_FILE = 2,
-    /* page that hold the page cache, for project 4 */
-    VM_PAGE_CACHE = 3,
-
-    /* Bit flags to store state */
-
-    /* Auxillary bit flag marker for store information. You can add more
-     * markers, until the value is fit in the int. */
-    VM_MARKER_0 = (1 << 3),
-    VM_MARKER_1 = (1 << 4),
-
-    /* DO NOT EXCEED THIS VALUE. */
-    VM_MARKER_END = (1 << 31),
-};
-
+#include "vm/uninit.h"
 #include "vm/anon.h"
 #include "vm/file.h"
-#include "vm/uninit.h"
+
 #ifdef EFILESYS
 #include "filesys/page_cache.h"
 #endif
@@ -59,6 +41,7 @@ struct page {
 #endif
     };
 
+    // 해시테이블 멤버 추가 08.04
     struct hash_elem hash_elem;
